Created BozemanTeleop and then GeminiTeleop. BozemanTeleop was an attempt to start to modify the robot components to reflect what we are planning to use for the Bozeman competition. I had Gemini help me reorganize and rewrite the code though and put that in GeminiTeleop. I think GeminiTeleop is much better and I will probably continue work on that in the future rather than the other TeleopFiles.
I also worked on the RRAutoActionTesting program and added motor actions - one that tells the motor to run to a position, and another that waits until the motor is done moving. That way you could be moving the arm while driving and completing other actions, but if there is a key moment where it matters that the arm is done moving and in position you can check for that before continuing with the routine. This combines the regular and patient action examples we had used in the past.